requirements.txt
Django==2.2.6
django-filter==2.2.0
djangorestframework==3.10.3
psycopg2==2.8.3